Dev ionut (#25)
* Update

* Update

* Update security

* Update jar

* Add method to remove AWS resources for user and its offers when user is deleted

* Add stripe and configure endpoint

* Add payment intent endpoint

* Add payment intent endpoint

* Add payment intent endpoint

* Update payment logic

* Increase request max size

* Update response for upload avatar

* Update delete account logic

* Add test

+++ src/main/java/com/auxby/usermanager/api/v1/user/UserService.java
@@ -3,9 +3,8 @@ package com.auxby.usermanager.api.v1.user;
 import com.auxby.usermanager.api.v1.address.model.AddressInfo;
 import com.auxby.usermanager.api.v1.auth.model.AuthInfo;
 import com.auxby.usermanager.api.v1.user.model.*;
-import com.auxby.usermanager.entity.Address;
-import com.auxby.usermanager.entity.Contact;
-import com.auxby.usermanager.entity.UserDetails;
+import com.auxby.usermanager.entity.*;
+import com.auxby.usermanager.exception.ActionNotAllowException;
 import com.auxby.usermanager.exception.ChangePasswordException;
 import com.auxby.usermanager.exception.RegistrationException;
 import com.auxby.usermanager.utils.enums.ContactType;
@@ -69,6 +68,14 @@ public class UserService {
     @Transactional
     public void deleteUser(String userUuid) {
         UserDetails userDetails = findUserDetails(userUuid);
+        List<Offer> onAuctionOffers = userDetails.getOffers()
+                .stream()
+                .filter(o -> o.isAvailable() && o.isOnAuction())
+                .toList();
+        var topBidders = userRepository.getTopBidderIdForOffers();
+        if (!onAuctionOffers.isEmpty() || topBidders.contains(userDetails.getId())) {
+            throw new ActionNotAllowException("Delete account not allow. User has active offers on auction or is top bidder for and offer.");
+        }
         keycloakService.deleteKeycloakUser(userDetails.getAccountUuid());
         deleteUserAwsResources(userUuid, userDetails);
         userRepository.deleteById(userDetails.getId());
